CINOIR pt.III

+++ database/migrations/2025_07_13_094912_add_status_to_movies_table.php
@@ -12,13 +12,25 @@ return new class extends Migration
     public function up()
 {
     Schema::table('movies', function (Blueprint $table) {
-        $table->enum('status', ['now_showing', 'coming_soon'])->default('now_showing')->after('durasi');
+        if (!Schema::hasColumn('movies', 'status')) {
+            $table->enum('status', ['now_showing', 'coming_soon'])->default('now_showing')->after('durasi');
+        }
+
+        if (!Schema::hasColumn('movies', 'rating')) {
+            $table->float('rating')->default(0)->after('status');
+        }
     });
 }
 public function down()
 {
     Schema::table('movies', function (Blueprint $table) {
-        $table->dropColumn('status');
+        if (Schema::hasColumn('movies', 'status')) {
+            $table->dropColumn('status');
+        }
+
+        if (Schema::hasColumn('movies', 'rating')) {
+            $table->dropColumn('rating');
+        }
     });
 }
 };
